Split JsonWriter into UTF-8 and UTF-16 specific types.
